Relocate AKS models
